Moved ObjectLike and DynamicMethods to common so Il2Cpp.Class could also benefit
